pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zerg-browser"
version = "0.1.0"
description = "Lightweight platform for managing and querying lab sequence data using natural language"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",

    # Data validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # File parsing
    "sgffp>=0.1",
    "biopython>=1.84",

    # Config
    "pyyaml>=6.0",

    # File watching
    "watchfiles>=1.0",

    # HTTP client (admin CLI)
    "httpx>=0.28",
]

[project.scripts]
zerg-admin = "zerg.admin.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "aiosqlite>=0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["src/zerg"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
